requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector search and AI
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.0
httpx<0.25.0
huggingface_hub==0.13.4


# Data processing
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6

# Image processing (for pipeline)
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# OCR and text extraction
pytesseract==0.3.10
easyocr==1.7.0

# File handling
python-magic==0.4.27
zipfile-deflate64==0.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pathlib2==2.3.7.post1

# Logging and monitoring
structlog==23.2.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development (optional)
black==23.11.0
flake8==6.1.0

# Flask (for backward compatibility with existing pipeline)
flask==3.0.0
flask-cors==4.0.0

# Additional ML libraries
scikit-learn==1.3.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.1+cpu
torchvision==0.16.1+cpu